chore: new commit message template
Often users don't know how to respect our commit guidelines, we loose
time explaining them how to write good commit messages. Using this
template the default commit message is changed to include a minimalistic
tutorial.

For now this template is opt-in, you can enable it via the following
command:

    git config commit.template .gitmessage

The "chore" type replaces the many "ci", "build", "test", "docs" and
"rel". It is the miscellaneous type to use when the change is not
related to the code.