feat: advanced logging configuration
Everything was logged with no way of silencing the logger, the result
was that the tool was very noisy.

We changed the base verbosity from INFO to WARNING and we added two cli
options to change it:

* -s/--silent: decrease verbosity
* -v/--verbose: increase verbosity

The two options can be repeated: -v: INFO, -vv DEBUG, -vvv DEBUG with
python warnings. -s: ERROR, -ss: CRITICAL (unused for now).

The logger used in `__main__.py` is one step more verbose than the rest
of the program. When the verbosity is configured on WARNING elsewhere,
it is INFO inside of `__main__.py`. This gives just enough messages when
used by regular users.

Closes: #22