feat(setuptools): specific var for python_requires
The PYTHON_VERSION is the python version currently in use, it does not
necessary correspond to the minimal python version that'll be supported
for the project. One can use py3.10 on his system (latest python as of
2022) but want his project to be compatible with py3.9 (as packaged in
debian bulleyes). This commit changes the variable to
AUDIENCE_PYTHON_VERSION which better reflect the purpose of the
variable.

Closes: #31