fix(venv): --upgrade-deps option missing on mint
The `--upgrade-deps` option is a recent addition to venv. This option
is missing from the python3-venv packages ship with the latest stable
Mint (Linux Mint Una 20.3) which cause incipyt to crash when
bootstraping a project.

It is safer to not use this option until it is deployed on all major
distributions.

Closes #27