feat(git): support update of user.email and user.name
Default values for AUTHOR_EMAIL and AUTHOR_NAME are retrieved with
git, test if the git command succeed before updating the
project.environ variables.

In case of undefined or uncompatible user.email and user.name, ask the
user if AUTHOR_EMAIL and AUTHOR_NAME should be used instead.

Add --force-vcs-config options to bypass user confirmation and force use
of AUTHOR_EMAIL and AUTHOR_NAME for user.email and user.name.