feat(license): choose a license from the CLI
A software license is a file stored at the root of a project which
explain what it is allowed to do with the project source code and under
what conditions. e.g. a project is qualified as "open-source" when it is
published with a license that allows to distribute and modify the
project source code.

Until now the license file generated by incipyt was a simple copyright
notice which made it explicit that the current work was the intellectual
property of its owner. Under internationnal laws, this forbid anyone
from modifying the work without the express authorization of the author.

In this work, we added a way so that the user can choose an open-source
license instead of the copyright notice. We added `--license <license>`
to the CLI arguments and bundled common open-source licenses such as MIT
and GPL.

The previous behavior, to simply include a copyright notice, is still
possible thanks to `--license copyright`.