feat: metadata system
In incipyt, it is difficult to express some technical needs such as to
define a variable as *required* only when some tools are enabled via the
CLI. It was also difficult to instinctively know in the source code
where the value of each variable was coming from, would the variable be
prompted to the user? In case we could via code determine a nice
default/final value, how could we inject it in the environment?

This work is a second take on a "metadata system", a new architecture
that should simplify the various variable metadata and that should make
it explicit where is variable value is coming from. See <#Prior work>
for a description of the first take.

Meta variables
--------------

A new module at `incipyt.project.meta_variables` stores the description
of every variable used within incipyt. All variables used in incipyt now
MUST be described inside that module. The description lists: the name of
the option, its type, a default suggestion/final value, whether that
variable is required (cannot be left empty), an help message to show in
`--help`.

All the variables are saved inside a mutable `variables` dict that is
importable via `incipyt.project`. Tools SHOULD NOT add new variables
inside that dict but they SHALL modify existing variables in case they
must change their definition. As example you can read how `poetry`
changes the `AUTHOR_NAME` and `AUTHOR_EMAIL` to make them required.

Environment
-----------

Prior to this commit, it was difficult to determine the original source
of every value saved inside the environment. It was a single dictionary
where all sources (cli/os.environ/runtime/hardcoded default/user prompt)
were mixed. At various places of the source code, new values were
injected in that dict and it was pretty difficult to reason about the
order in which every value was injected and whether the user would be
prompted to use any of those.

This commit completely changes the approach to this problem. All the
values of a same source and now stored in dedicated dictionary (actually
two dictionaries, one hold confirmed values, values for which the user
will not be prompted and a dictionary for unconfirmed values for which
he will upon first access). The result is a wompy 7 dictionaries, 2 for
source hardcoded values, 2 for os.environ, 2 for runtime (tools)
determined values and 1 for the values injected via the CLI (no
unconfirmed dictionary for this source). The actual injection of
variable values occurs very early at the beginning of `__main__` and the
`environ` dict starts empty.

Lookup order
------------

Upon variable access, the following algorithm is used to determine the
value.

1. It uses the value present in `environ` if it is set.
2. It lookups the many sources to find the variable, it will prompt a
   confirmation to the user in case it finds the value in an unconfirmed
   source.
3. In case the lookup fails, i.e. that no source has a value set for
   that variable, the user is prompted. In case the variable is
   required, it will prompt the user until he gives a non-empty value.

Value determined in (2) and (3) are saved in `environ`.

The lookup order is as follow, it prompts the user for a confirmation
starting at 5.²

1. CLI -o/--option (new!)
2. Runtime `environ.inject()`-ed value
3. `os.environ` if the variable starts with `INCIPYT_`
4. Hardcoded `prompt=False` default value
5. Runtime `environ.suggest()`-ed value
6. Other `os.environ`
7. Hardcoded `prompt=True` default value

Changes to the API
------------------

`environ.inject(var, value)` and `environ.suggest(var, value)` are two
new methods on `project.environ`. They replace the previous ways of
setting programmatically a variable inside of the environment.

* `environ.inject(var, value)` replaces `environ[var] = value`
* `environ.suggest(var, value)` replace `environ.setdefault(var, value)`

It is still possible to `__setitem__`/`setdefault` but is discouraged as
this would bypass the new variable cascading lookup.

Prior work
----------

Florian was the first to implement metadata system based on various
architectural discussions we had. He wrote the first version of the
`Variable` class (called `_EnvMetadata` in his work), a dataclass with
all the meta-information on every variable: name, type, default,
require, etc. He also had a few functions dedicated at modifying those
variables (the dict where variables were stored was private in his
work). His work lacked a definitive solution on how to load the many
sources in a sensitive way and there was still some rooms for
improvements.

My (Julien) contribution was the `ChainMap` where every source is loaded
in separate dictionaries. I also embraced the variable dataclass way and
pushed the refactoring even further.

Both works have been merged as mine is based on his. His work is visible
at: c0e32c464dc4e102edd44f016f91ed76a445b1fc..e9f784c029f73864811601441c37e091d0114491