feat: config file
Users don't want to provide all -o/--option values to configure how
their new project will be generated, they want to benefit from incipyt
wizard and be prompted sensitive default values. At the same time, the
various default values used by incipyt are not always the best ones.

We added a new toml configuration file that can be modified by users to
customize their experience with incipyt. The configuration file is
stored at a sensitive place depending on the operating system (see
below) and is a very simple `VARIABLE=value` format.

The available variables are the same as those listed in `--help`.

The default location for the configuration is listed below, it can be
override by providing the `-c` or `--config` argument on the command
line.

* Windows: C:\Users\<user>\AppData\Roaming\incipyt\config.toml
* Mac OS: ~/Library/Application Support/incipyt/config.toml
* Linux: ~/.config/incipyt/config.toml